Refactor Scan tests
- Move test description comments to docstrings
- Move optimization tests to `test_opt`
- Move `aesara.scan.views` tests to a new `test_views`
- Move example-based tests (i.e. mostly non-functionality-focused tests) to a
  distinct test suite/class
- Make rewrite tests specify the presence of the relevant rewrites (assuming
  that those tests even rely on their namesake rewrites, of course)
- Miscellaneous test updates
